pufffff

+++ CentiroHomeAssignment/DTOs/OrderDTO.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace CentiroHomeAssignment.DTOs
 {
@@ -16,7 +13,7 @@ namespace CentiroHomeAssignment.DTOs
         public string ProductNumber { get; set; }
         public string ProductName { get; set; }
         public string Description { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public string ProductGroup { get; set; }
     }
 }
